requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.1
xgboost==2.1.1
imbalanced-learn==0.12.3
shap==0.46.0
reportlab==4.2.2
plotly==5.24.1
joblib==1.4.2

Jinja2==3.1.4
